requirements.txt
Django==1.6
dj-database-url==0.3.0
dj-static==0.0.5
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.3
pystache==0.5.4
static==1.0.2
wsgiref==0.1.2
PyYAML==3.11
South==0.8
django-bootstrap3-datetimepicker==2.2.3
django-guardian==1.2.0
django-registration==1.0
django-userena==1.3.0
easy-thumbnails==2.0.1

